# docs/04-api-contract.yaml
openapi: 3.0.3
info:
  title: My-AI API
  version: 0.1.0
  description: |
    V1.1 接口契约（知识库主数据化 + 文档列表管理基线）

    RAG 权限体系阶段二开始后，除登录接口和健康检查外，`/api/v1/**` 默认需要
    `HttpOnly Cookie` 中的服务端 Session。所有写操作需要携带固定 Header：
    `X-MYAI-CSRF: 1`。
servers:
  - url: http://localhost:8080
...
paths:
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: 本地账号登录
      description: |
        登录成功后创建服务端 Session，并通过 HttpOnly Cookie 维持认证态。
        当前 CSRF 基线要求该写操作携带 `X-MYAI-CSRF: 1`。
      x-stage: implemented
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: 请求体为空或参数缺失
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 用户名或密码错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 缺少 CSRF Header、账号停用或账号锁定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: 登出当前 Session
      x-stage: implemented
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '204':
          description: 登出成功
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 缺少或非法 CSRF Header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: 查询当前登录用户
      description: |
        返回当前登录用户的基础身份信息与前端菜单渲染所需的能力位。

        当前能力位采用嵌套对象 `capabilities` 返回，前端据此决定：
        - 一级菜单是否显示
        - 登录后的默认落点
        - 业务页无权限时跳转到哪个可访问页面
      x-stage: implemented
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/audit-events:
    get:
      tags: [governance]
      summary: 查询审计事件分页列表
      description: |
        返回当前工作区内的审计事件分页结果；
        默认按 `occurredAt DESC, auditEventId DESC` 排序；
        支持按事件类型、操作者、目标类型、目标 ID、结果代码与时间范围过滤；
        仅工作区所有者或管理员可访问。
      x-stage: implemented
      security:
        - cookieAuth: []
      parameters:
        - name: eventType
          in: query
          required: false
          schema:
            type: string
          description: 事件类型过滤，如 `LOGIN_SUCCESS`、`DOCUMENT_GRANT_UPSERTED`
        - name: actorUserId
          in: query
          required: false
          schema:
            type: string
          description: 操作者用户 ID 过滤
        - name: actorKeyword
          in: query
          required: false
          schema:
            type: string
          description: 操作者关键词过滤，支持用户名或用户 ID 模糊匹配
        - name: targetType
          in: query
          required: false
          schema:
            type: string
          description: 目标类型过滤，如 `USER`、`WORKSPACE_MEMBERSHIP`、`KNOWLEDGE_BASE_GRANT`、`DOCUMENT_GRANT`
        - name: targetId
          in: query
          required: false
          schema:
            type: string
          description: 目标 ID 过滤
        - name: outcome
          in: query
          required: false
          schema:
            type: string
            enum: [SUCCESS, FAILURE, DENIED]
          description: 结果代码过滤
        - name: occurredFrom
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: 起始时间（含）
        - name: occurredTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: 结束时间（含）
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: 每页条数
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: 分页偏移量
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventPageResponse'
        '400':
          description: 查询参数非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 当前用户无审计查询权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
    CsrfHeader:
      name: X-MYAI-CSRF
      in: header
      required: true
      schema:
        type: string
        enum: ['1']
      description: 写操作统一 CSRF Header，当前固定值为 `1`
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/CurrentUserResponse'
    CurrentUserResponse:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        displayName:
          type: string
        workspaceId:
          type: string
          example: default
        workspaceRole:
          type: string
          enum: [WORKSPACE_OWNER, WORKSPACE_ADMIN, WORKSPACE_MEMBER]
        capabilities:
          $ref: '#/components/schemas/CurrentUserCapabilitiesResponse'
    CurrentUserCapabilitiesResponse:
      type: object
      properties:
        canAccessDocumentList:
          type: boolean
        canUploadDocument:
          type: boolean
        canAccessKnowledge:
          type: boolean
        canAskQuestion:
          type: boolean
        canAccessAdmin:
          type: boolean
...
    AuditEventPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    AuditEvent:
      type: object
      properties:
        auditEventId:
          type: integer
          format: int64
        workspaceId:
          type: string
          nullable: true
        actorUserId:
          type: string
          nullable: true
        actorUsername:
          type: string
          nullable: true
        eventType:
          type: string
        targetType:
          type: string
          nullable: true
        targetId:
          type: string
          nullable: true
        outcome:
          type: string
          enum: [SUCCESS, FAILURE, DENIED]
        reason:
          type: string
        metadata:
          type: object
          additionalProperties: true
        occurredAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...